requirements.txt
absl-py==1.2.0
accelerate==0.21.0
aiohttp==3.8.2
aiosignal==1.2.0
albumentations==1.3.0
ale-py==0.7.4
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
antlr4-python3-runtime==4.9.3
array-record==0.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autofaiss==2.15.5
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
blis==0.7.10
box2d-py==2.3.5
braceexpand==0.1.7
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
catalogue==2.0.9
catboost==1.1.1
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1606255122322/work
chardet @ file:///tmp/build/80754af9/chardet_1605303192865/work
charset-normalizer==2.1.1
chex==0.1.7
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cloudpickle==2.2.1
clu==0.0.9
cmake==3.26.3
colorama==0.4.6
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1605484832764/work
confection==0.1.1
contextlib2==21.6.0
contourpy==1.0.7
crcmod==1.7
cryptography @ file:///tmp/build/80754af9/cryptography_1607635322643/work
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
dask==2023.4.0
dataclasses==0.6
datasets==2.12.0
deepspeed==0.9.2
dill==0.3.6
distributed==2023.4.0
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
dotmap==1.3.30
embedding-reader==1.5.0
etils==1.2.0
ExifRead-nocycle==3.0.1
faiss-gpu==1.7.2
Farama-Notifications==0.0.4
filelock==3.8.0
fire==0.4.0
flatbuffers==23.3.3
flax==0.6.9
fonttools==4.39.3
frozenlist==1.3.1
fsspec==2022.11.0
ftfy==6.1.1
fuzzywuzzy==0.18.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
glcontext==2.3.7
glfw==1.12.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
grpcio==1.53.0
gym==0.21.0
gym3==0.3.3
gymnasium==0.28.1
h5py==3.8.0
HeapDict==1.0.1
hjson==3.1.0
huggingface-hub==0.15.1
hydra-core==1.3.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.27.0
imageio-ffmpeg==0.3.0
img2dataset==1.41.0
immutabledict==2.2.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
jax==0.4.8
jax-jumpy==1.0.0
jaxlib==0.4.7+cuda11.cudnn86
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonpickle==3.0.1
kaggle==1.5.13
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy-loader==0.2
libclang==16.0.0
lightgbm==3.3.5
lit==16.0.5
llvmlite==0.39.1
locket==1.0.0
lvis==0.5.3
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.1.0
moderngl==5.8.2
mpmath==1.3.0
msgpack==1.0.5
multidict==5.2.0
multiprocess==0.70.14
munch==2.5.0
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
ninja==1.10.2.4
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.1
omegaconf==2.3.0
open-clip-torch==2.16.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opendatalab==0.0.10
openxlab==0.0.22
opt-einsum==3.3.0
optax @ git+https://github.com/deepmind/optax.git@e5f910e92534c3b444e806582227389b7bb2b3b0
orbax-checkpoint==0.2.1
oss2==2.17.0
ott-jax==0.3.1
packaging==21.3
pandas==1.5.1
partd==1.4.0
pathtools==0.1.2
pathy==0.10.2
Pillow==7.2.0
pipreqs==0.4.13
plotly==5.14.1
preshed==3.0.8
progressbar2==4.2.0
promise==2.3
protobuf==3.20.3
psutil==5.9.2
py-cpuinfo==8.0.0
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pybullet==3.2.5
pycocotools==2.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.18.0
pydantic==1.10.2
pyDeprecate==0.3.2
pyglet==2.0.5
Pygments==2.15.1
pynverse==0.1.4.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1606517880428/work
pyparsing==3.0.9
pypoly2tri==0.0.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-slugify==8.0.1
python-utils==3.6.0
pytorch-lightning==1.7.6
pytz==2023.3
PyVirtualDisplay==3.0
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2022.9.13
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.4.2
rsa==4.9
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1605527293619/work
sacred==0.8.4
safetensors==0.3.1
scenic @ file:///work/itr176/Research_Truman_Hickok/optimal_rehearsal/scenic
scikit-image==0.20.0
scikit-learn==1.1.3
scipy==1.9.1
sentencepiece==0.1.97
sentry-sdk==1.9.8
setproctitle==1.3.2
Shapely==1.8.5.post1
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1605205306277/work
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
sortedcontainers==2.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stable-baselines3==1.8.0
sympy==1.12
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
tensorstore==0.1.36
termcolor==2.2.0
text-unidecode==1.3
thinc==8.1.11
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.11
tokenizers==0.13.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0+cu117
torch-summary==1.4.5
torchaudio==2.0.1+cu117
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.15.1+cu117
tornado==6.2
tqdm==4.65.2
transformers==4.31.0
triton==2.0.0
typeguard==2.13.3
typer==0.9.0
typing-extensions==4.3.0
urllib3==1.26.12
visual-genome==1.1.1
wandb==0.12.21
wasabi==1.1.2
wcwidth==0.2.6
webdataset==0.2.48
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.7.5
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.1
zict==2.2.0
zipp==3.8.1
